backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Background worker
rq==1.16.2
redis==5.2.1

# HTTP client (for Nano Banana Pro + Gemini calls)
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Utils
python-multipart==0.0.20
Pillow==11.1.0
